add calculateCurrentPosition, update tests
